requirements.txt
sanic==20.9.1
Sanic-Cors==0.10.0.post3
xmltodict==0.12.0
python-gvm